refector(*): 우승 결과를 구하는 책임을 WinningLotto가 가지도록 수정
