+++ src/main/java/domain/WinningLotto.java
@@ -32,4 +32,15 @@ public class WinningLotto {
             throw new IllegalArgumentException("당첨 번호와 보너스 번호는 중복될 수 없습니다.");
         }
     }
+
+    public WinningResult calculateWinningResult(LottoWallet lottoWallet) {
+        WinningResult winningResult = new WinningResult();
+        for (Lotto lotto : lottoWallet.getLottoWallet()) {
+            final int matchedCount = countMatchedNumber(lotto);
+            final boolean isBonusMatched = isBonusNumberMatched(lotto);
+            WinningInfo winningInfo = WinningInfo.of(matchedCount, isBonusMatched);
+            winningResult.increaseCount(winningInfo, 1);
+        }
+        return winningResult;
+    }
 }
