+++ src/test/java/LottoManagerTest.java
@@ -16,15 +16,17 @@ import org.junit.jupiter.api.Test;
 
 class LottoManagerTest {
     private LottoManager lottoManager;
-    List<Lotto> lottos = List.of(
+    private final List<Lotto> lottos = List.of(
             new Lotto(List.of(1, 2, 3, 4, 5, 6)),
             new Lotto(List.of(1, 2, 3, 4, 5, 7)),
             new Lotto(List.of(1, 2, 3, 4, 5, 8)),
             new Lotto(List.of(1, 2, 3, 4, 10, 11)),
             new Lotto(List.of(1, 2, 3, 10, 11, 12)),
             new Lotto(List.of(10, 11, 12, 13, 14, 15)));
-    LottoWallet lottoWallet = new LottoWallet(lottos);
-    WinningLotto winningLotto = new WinningLotto(new Lotto(List.of(1, 2, 3, 4, 5, 6)), new LottoNumber(7));
+    private final LottoWallet lottoWallet = new LottoWallet(lottos);
+    private final WinningLotto winningLotto = new WinningLotto(new Lotto(List.of(1, 2, 3, 4, 5, 6)),
+            new LottoNumber(7));
+
 
     @BeforeEach
     void setup() {
@@ -45,26 +47,10 @@ class LottoManagerTest {
     }
 
     @Test
-    @DisplayName("로또와 당첨 번호가 주어졌을 때 로또 순위별 당첨 횟수를 정확히 계산한다")
-    void check_each_winning_count_correctly() {
-        // given
-        LottoManager lottoManager = new LottoManager();
-
-        // when
-        WinningResult winningResult = lottoManager.calculateWinningResult(lottoWallet, winningLotto);
-
-        // then
-        for (WinningInfo value : WinningInfo.values()) {
-            assertThat(winningResult.getCount(value)).isEqualTo(1);
-        }
-    }
-
-    @Test
     @DisplayName("구매 금액과 로또 결과가 주어졌을 때 수익률을 정확히 계산한다")
     void check_calculate_revenue_correctly() {
         // given
-        LottoManager lottoManager = new LottoManager();
-        WinningResult winningResult = lottoManager.calculateWinningResult(lottoWallet, winningLotto);
+        WinningResult winningResult = winningLotto.calculateWinningResult(lottoWallet);
 
         Money money = new Money(6000);
         long totalPrices = Arrays.stream(WinningInfo.values())
@@ -78,5 +64,4 @@ class LottoManagerTest {
         // then
         assertThat(result).isEqualTo(expected);
     }
-
 }
