+++ src/test/java/WinningLottoTest.java
@@ -4,7 +4,12 @@ import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 import domain.Lotto;
 import domain.LottoNumber;
+import domain.LottoWallet;
+import domain.Money;
+import domain.WinningInfo;
 import domain.WinningLotto;
+import domain.WinningResult;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
@@ -15,6 +20,15 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
 class WinningLottoTest {
+    private List<Lotto> lottos = List.of(
+            new Lotto(List.of(1, 2, 3, 4, 5, 6)),
+            new Lotto(List.of(1, 2, 3, 4, 5, 7)),
+            new Lotto(List.of(1, 2, 3, 4, 5, 8)),
+            new Lotto(List.of(1, 2, 3, 4, 10, 11)),
+            new Lotto(List.of(1, 2, 3, 10, 11, 12)),
+            new Lotto(List.of(10, 11, 12, 13, 14, 15)));
+    private LottoWallet lottoWallet = new LottoWallet(lottos);
+    private WinningLotto winningLotto = new WinningLotto(new Lotto(List.of(1, 2, 3, 4, 5, 6)), new LottoNumber(7));
 
     @Test
     @DisplayName("당첨 번호에 보너스 번호가 존재하면 예외가 발생한다")
@@ -64,5 +78,18 @@ class WinningLottoTest {
 
         // then
         assertThat(isBonusMatched).isEqualTo(expected);
+    }
+
+
+    @Test
+    @DisplayName("로또와 당첨 번호가 주어졌을 때 로또 순위별 당첨 횟수를 정확히 계산한다")
+    void check_each_winning_count_correctly() {
+        // when
+        WinningResult winningResult = winningLotto.calculateWinningResult(lottoWallet);
+
+        // then
+        for (WinningInfo value : WinningInfo.values()) {
+            assertThat(winningResult.getCount(value)).isEqualTo(1);
+        }
     }
 }
